NOTICE
Dear Diskette Theme Customer,

Please find enclosed the patch or update disk that you
requested'. 

For full instructions on updating your theme.css, you
can view the README file on the disk. 

To view this file:

1) Insert the disk into your disk drive and make this drive
your current directory.

For example, if you put the disk in
your A: drive, type A:

FIX
or
README

You can also view the README file with a text editor or
word processor.